flow_builder_agent/requirements.txt
# Flow Builder Agent Requirements

# Core dependencies
pydantic>=2.0.0
openai>=1.0.0
google-generativeai>=0.3.0
requests>=2.25.0

# RAG and ML dependencies  
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
numpy>=1.21.0

# Utility dependencies
python-dotenv>=1.0.0

# Development dependencies (optional)
pytest>=7.0.0
black>=22.0.0
isort>=5.0.0
mypy>=1.0.0
